[Cases] Reduce amount of sync tasks by a factor of 4 (#238276)
## Summary

This PR reduces the effective amount of tasks used for synchronizing the
cases analytics index from 12 per space to 3 per space. In larger Kibana
deployments, this should result in a lot less stress on the Task
Manager.

Instead of scheduling a task for each combination of space/owner/sync
type, we only schedule a task per space/owner. This is done by moving
the different sync types into a single task which then starts "sub
tasks" per sync type. These are not real tasks, just ephemeral
asynchronous functions.

```mermaid
flowchart TD
    A(Synchronization Task)
    B(Synchronization Task)
    D(Synchronization Task)
    E(Synchronization Task)
    C("(Observability)")
    A1("Scheduler Task")

    subgraph v1
        A1-->C
        A1-->C2
        A1-->C3
        EC@{ shape: braces, label: "state passed:
        sourceIndex
        destIndex
        owner
        spaceId
        syncType" }
        C-->A
        C-->B
        C-->D
        C-->E
        C2-->A2(...)
        C2-->B2(...)
        C2-->D2(...)
        C2-->E2(...)
        C3-->A3(...)
        C3-->B3(...)
        C3-->D3(...)
        C3-->E3(...)
        C2("(Security)")
        C3("(Stack)")
    end

    subgraph v2
        X("Scheduler Task")
        Y("(Observability)")
        Y2("(Security)")
        Y3("(Stack)")
        Z("Synchronization Task")
        Z2("Synchronization Task")
        Z3("Synchronization Task")

        X-->Y
        X-->Y2
        X-->Y3

        Y-->Z
        Y2-->Z2
        Y3-->Z3
        EC2@{ shape: braces, label: "state passed:
        owner
        spaceId" }
    end
```

### Testing

See `Testing` section in original PR:
https://github.com/elastic/kibana/pull/234125

(cherry picked from commit d7d7c927eea0fd136146a93671c88bfa5e235146)
